+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 
 - Explicitly rescue NameError in Cache#presenter_classes_for
+- Add Oprah::TestHelpers
 
 0.1.1
 -----